Move ffi stream and utils from arrow to arrow-array
